Move API v1 tests to separate tree
